making build process multithreaded; adding physics scenes to build

+++ tasks.py
@@ -2,6 +2,7 @@ import pprint
 import os
 import datetime
 import zipfile
+import threading
 import hashlib
 import subprocess
 from invoke import task
@@ -12,8 +13,8 @@ def add_files(zipf, start_dir):
     for root, dirs, files in os.walk(start_dir):
         for f in files:
             fn = os.path.join(root, f)
-            arcname = os.path.relpath(fn, 'unity/builds')
-            #print("adding %s" % arcname)
+            arcname = os.path.relpath(fn, start_dir)
+            # print("adding %s" % arcname)
             zipf.write(fn, arcname)
 
 def push_build(build_archive_name):
@@ -27,18 +28,19 @@ def push_build(build_archive_name):
     print("pushed build %s to %s" % (S3_BUCKET, build_archive_name))
 
 
-def _build(context, arch, build_name):
+def _build(context, arch, build_dir, build_name):
     project_path = os.path.join(os.getcwd(), 'unity')
     command = "/Applications/Unity-2017.3.1f1/Unity.app/Contents/MacOS/Unity -quit -batchmode -logFile build.log -projectpath %s -executeMethod Build.%s" % (project_path, arch)
+    target_path = os.path.join(build_dir, build_name)
 
-    return context.run(command, warn=True, env=dict(UNITY_BUILD_NAME=build_name))
+    return context.run(command, warn=True, env=dict(UNITY_BUILD_NAME=target_path))
 
 @task
 def local_build(context, prefix='local'):
     arch = 'OSXIntel64'
-    build_name = "builds/thor-%s-%s" % (prefix, arch)
+    build_name = "thor-%s-%s" % (prefix, arch)
     fetch_source_textures(context)
-    if _build(context, arch, build_name):
+    if _build(context, arch, "builds", build_name):
         print("Build Successful")
     else:
         print("Build Failure")
@@ -92,7 +94,7 @@ def build_sha256(path):
 def build_docker(version):
 
     subprocess.check_call(
-        "docker build --rm --no-cache --build-arg AI2THOR_VERSION={version} -t  ai2thor/ai2thor-base:{version} .".format(version=version),
+        "docker build --quiet --rm --no-cache -t  ai2thor/ai2thor-base:{version} .".format(version=version),
         shell=True)
 
     subprocess.check_call(
@@ -103,7 +105,10 @@ def build_docker(version):
 def build_pip(context):
     import shutil
     subprocess.check_call("python setup.py clean --all", shell=True)
-    shutil.rmtree("dist")
+
+    if os.path.isdir('dist'):
+        shutil.rmtree("dist")
+
     subprocess.check_call("python setup.py sdist bdist_wheel --universal", shell=True)
 
 @task
@@ -118,39 +123,59 @@ def fetch_source_textures(context):
     z = zipfile.ZipFile(io.BytesIO(zip_data))
     z.extractall(os.getcwd())
 
+def archive_push(build_path, build_dir, build_info):
+    threading.current_thread().success = False
+    archive_name = os.path.join('unity', build_path)
+    zipf = zipfile.ZipFile(archive_name, 'w', zipfile.ZIP_STORED)
+    add_files(zipf, os.path.join('unity', build_dir))
+    zipf.close()
+
+    build_info['sha256'] = build_sha256(archive_name)
+    push_build(archive_name)
+    print("Build successful")
+    threading.current_thread().success = True
+
 @task
 def build(context, local=False):
+    from multiprocessing import Process
     version = datetime.datetime.now().strftime('%Y%m%d%H%M')
     build_url_base = 'http://s3-us-west-2.amazonaws.com/%s/' % S3_BUCKET
 
     builds = {'Docker': {'tag': version}}
     fetch_source_textures(context)
+    threads = []
+    dp = Process(target=build_docker, args=(version,))
+    dp.start()
 
-    for arch in ['OSXIntel64', 'Linux64']:
-        build_name = "builds/thor-%s-%s" % (version, arch)
-        url = build_url_base + build_name + ".zip"
+    #for arch in ['OSXIntel64']:
+    platform_map = dict(Linux64="Linux", OSXIntel64="Darwin")
 
-        x = _build(context, arch, build_name)
+    for arch in platform_map.keys():
+        build_name = "thor-%s-%s" % (version, arch)
+        build_dir = os.path.join('builds', build_name)
+        build_path = build_dir + ".zip"
+        build_info = builds[platform_map[arch]] = {}
 
-        build_info = None
+        build_info['url'] = build_url_base + build_path
+
+        x = _build(context, arch, build_dir, build_name)
+
         if x:
-            archive_name = os.path.join('unity', build_name + ".zip")
-            zipf = zipfile.ZipFile(archive_name, 'w', zipfile.ZIP_DEFLATED)
-            if arch == 'OSXIntel64':
-                add_files(zipf, os.path.join('unity', build_name + ".app"))
-                build_info = builds['Darwin'] = {}
-            elif arch == 'Linux64':
-                build_info = builds['Linux'] = {}
-                add_files(zipf, os.path.join('unity', build_name + "_Data"))
-                zipf.write(os.path.join('unity', build_name), os.path.basename(build_name))
-            build_info['url'] = url
-            zipf.close()
-
-            build_info['sha256'] = build_sha256(archive_name)
-            push_build(archive_name)
-            print("Build successful")
+            t = threading.Thread(target=archive_push, args=(build_path, build_dir, build_info))
+            t.start()
+            threads.append(t)
         else:
             raise Exception("Build Failure")
+
+    dp.join()
+
+    if dp.exitcode != 0:
+        raise Exception("Exception with docker build")
+
+    for t in threads:
+        t.join()
+        if not t.success:
+            raise Exception("Error with thread")
 
     generate_quality_settings(context)
 
@@ -160,7 +185,6 @@ def build(context, local=False):
         fi.write("BUILDS = " + pprint.pformat(builds))
 
     increment_version()
-    build_docker(version)
     build_pip(context)
 
 @task
